Update README.md
This is a console application which is developed in C - language. This application is simply a game in which we have to sort the given 4 by 4 matrix without violating the rules of the game that will be visible us before game start.

Developed a console application name's Number Shifting Game which sort's a matrix by shifting number one by one. By playing this game it's develop curiosity and excitement, how actually it's happening and how to do it.

In this game by shifting number on empty space we can sort matrix in limited number of moves, if player does this job that means he/she won the game else loose.

While playing game user interacting with application by pressing only arrow keys to move number one by one on empty space to sort matrix.

It's really develops curiosity and excitement how we can also do this  using C language.